advanced_agents/llm-sql-agent/requirements.txt
langchain
langgraph
langchain-community
langchain-google-genai
google-generativeai
sqlalchemy
python-dotenv
pydantic
pyyaml
pytest
pytest-asyncio
colorama
tabulate